+++ cds-plugin.js
@@ -12,10 +12,7 @@ if (cds.env.cov2ap.plugin) {
   cds.on("bootstrap", async (app) => {
     app.use(cov2ap.singleton());
   });
-  if (cds.env.cov2ap.build && cds.build && cds.build.register) {
-    cds.build.register("cov2ap", {
-      impl: "@cap-js-community/odata-v2-adapter/src/build.js",
-      taskDefaults: { src: "srv" },
-    });
+  if (cds.env.cov2ap.build) {
+    cds.build?.register("cov2ap", require("./src/build"));
   }
 }